MA-22366-4 Camera: Fix coverity issue in image_process
this fix below coverity issues:
CID 38744426: Macro compares unsigned to 0
CID 34684784: Uninitialized scalar variable (UNINIT)

Reviewed-on: http://androidsource.nxp.com/project/25362

+++ image_process/ImageUtils.cpp
@@ -554,7 +554,7 @@ void Rgb888ToYuv422i(const uint8_t *rgb, uint8_t *yuv422i, int width, int height
 
     uint16_t i, j;
     uint8_t r, g, b;
-    uint8_t y, u, v;
+    int y, u, v;
 
     for (i = 0; i < height; i++) {
         for (j = 0; j < width; j++) {
